# specs/001-create-an-api/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Habit Tracker API
  description: API-first habit tracking system with visual progress grids
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://your-app.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - ApiKeyAuth: []
paths:
...
  /auth/verify:
    post:
      summary: Verify magic link token
      description: Completes authentication and returns API key
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Magic link token from email
              required:
                - token
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  api_key:
                    type: string
                    description: Generated API key for future requests
        '401':
          $ref: '#/components/responses/Unauthorized'

  # API Key Management
...
  /boards:
    get:
      summary: List user's boards
      description: Get all habit boards for authenticated user
      parameters:
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
          description: Include archived boards
      responses:
        '200':
          description: List of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'

    post:
      summary: Create new board
      description: Create a new habit tracking board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Morning Exercise"
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                  default: '#22c55e'
                icon:
                  type: string
                  maxLength: 10
              required:
                - name
      responses:
        '201':
          description: Board created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Use format: Bearer <api_key>'
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Board:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Morning Exercise"
        description:
          type: string
          nullable: true
        color:
          type: string
          example: "#22c55e"
        icon:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            code: "VALIDATION_ERROR"
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid API key"
            code: "UNAUTHORIZED"
...